handle existing symlinks

+++ bin/desi_spectro_data_move.sh
@@ -72,10 +72,18 @@ release_data=${DESI_ROOT}/public/${release}/spectro/data
 relative_data='../../public/edr/spectro/data'
 for n in ${DESI_SPECTRO_DATA}/20*; do
     night=$(basename ${n})
-    if is_night_in_release ${release} ${night}; then
-        ${verbose} && echo "DEBUG: mv -v ${DESI_SPECTRO_DATA}/${night} ${release_data}"
-        ${test}    || mv -v ${DESI_SPECTRO_DATA}/${night} ${release_data}
-        ${verbose} && echo "DEBUG: (cd ${DESI_SPECTRO_DATA} && ln -s -v ${relative_data}/${night})"
-        ${test}    || (cd ${DESI_SPECTRO_DATA} && ln -s -v ${relative_data}/${night})
+    if [[ -L ${n} ]]; then
+        echo "INFO: ${n} is already a symlink."
+    else
+        if is_night_in_release ${release} ${night}; then
+            ${verbose} && echo "DEBUG: chmod -v u+w ${DESI_SPECTRO_DATA}/${night}"
+            ${test}    || chmod -v u+w ${DESI_SPECTRO_DATA}/${night}
+            ${verbose} && echo "DEBUG: mv -v ${DESI_SPECTRO_DATA}/${night} ${release_data}"
+            ${test}    || mv -v ${DESI_SPECTRO_DATA}/${night} ${release_data}
+            ${verbose} && echo "DEBUG: chmod -v u-w ${release_data}/${night}"
+            ${test}    || chmod -v u-w ${release_data}/${night}
+            ${verbose} && echo "DEBUG: (cd ${DESI_SPECTRO_DATA} && ln -s -v ${relative_data}/${night})"
+            ${test}    || (cd ${DESI_SPECTRO_DATA} && ln -s -v ${relative_data}/${night})
+        fi
     fi
 done
